blackList/requirements.txt
Django==5.1
djangorestframework==3.15.2
requests==2.32.3
django-cors-headers==4.4.0
psycopg2==2.9.9